requirements.txt
django>3.2,<3.3
python-decouple
Pillow
django-simple-math-captcha==1.1.0
django-debug-toolbar
django-ckeditor
djangorestframework
markdown
django-filter
coreapi
drf-yasg[validation]
djangorestframework-simplejwt==4.7.2
djoser
django-mail-templated
black
flake8
pytest
pytest-django